Refactor into utils module
