+++ nucliadb/nucliadb/export_import/exporter.py
@@ -18,18 +18,22 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 
-from typing import AsyncGenerator, AsyncIterator
+from typing import AsyncGenerator
 
 from nucliadb.common.context import ApplicationContext
-from nucliadb.common.datamanagers.entities import EntitiesDataManager
-from nucliadb.common.datamanagers.labels import LabelsDataManager
-from nucliadb.common.datamanagers.resources import ResourcesDataManager
 from nucliadb.export_import.models import ExportedItemType
-from nucliadb_protos import knowledgebox_pb2 as kb_pb2
-from nucliadb_protos import resources_pb2, writer_pb2
+from nucliadb.export_import.utils import (
+    download_binary,
+    get_binaries,
+    get_entities,
+    get_labels,
+    iter_broker_messages,
+)
 
 
-async def export_kb(context: ApplicationContext, kbid: str) -> AsyncIterator[bytes]:
+async def export_kb(
+    context: ApplicationContext, kbid: str
+) -> AsyncGenerator[bytes, None]:
     """Export the data of a knowledgebox to a stream of bytes.
 
     See https://github.com/nuclia/nucliadb/blob/main/docs/internal/EXPORTS.md
@@ -67,86 +71,3 @@ async def export_kb(context: ApplicationContext, kbid: str) -> AsyncIterator[byt
         yield ExportedItemType.LABELS.encode("utf-8")
         yield len(bytes).to_bytes(4, byteorder="big")
         yield bytes
-
-
-async def iter_broker_messages(
-    context: ApplicationContext, kbid: str
-) -> AsyncGenerator[writer_pb2.BrokerMessage, None]:
-    rdm = ResourcesDataManager(context.kv_driver, context.blob_storage)
-    async for rid in rdm.iterate_resource_ids(kbid):
-        resource = await rdm.get_resource(kbid, rid)
-        if resource is None:
-            continue
-        resource.disable_vectors = False
-        bm = await resource.generate_broker_message()
-        yield bm
-
-
-def get_binaries(bm: writer_pb2.BrokerMessage) -> list[resources_pb2.CloudFile]:
-    """Return the list of binaries of a broker message."""
-    binaries: list[resources_pb2.CloudFile] = []
-    for file_field in bm.files.values():
-        if file_field.HasField("file"):
-            _clone_collect_cf(binaries, file_field.file)
-
-    for conversation in bm.conversations.values():
-        for message in conversation.messages:
-            for attachment in message.content.attachments:
-                _clone_collect_cf(binaries, attachment)
-
-    for layout in bm.layouts.values():
-        for block in layout.body.blocks.values():
-            if block.HasField("file"):
-                _clone_collect_cf(binaries, block.file)
-
-    for field_extracted_data in bm.file_extracted_data:
-        if field_extracted_data.HasField("file_thumbnail"):
-            _clone_collect_cf(binaries, field_extracted_data.file_thumbnail)
-        if field_extracted_data.HasField("file_preview"):
-            _clone_collect_cf(binaries, field_extracted_data.file_preview)
-        for file_generated in field_extracted_data.file_generated.values():
-            _clone_collect_cf(binaries, file_generated)
-        for page in field_extracted_data.file_pages_previews.pages:
-            _clone_collect_cf(binaries, page)
-
-    for link_extracted_data in bm.link_extracted_data:
-        if link_extracted_data.HasField("link_thumbnail"):
-            _clone_collect_cf(binaries, link_extracted_data.link_thumbnail)
-        if link_extracted_data.HasField("link_preview"):
-            _clone_collect_cf(binaries, link_extracted_data.link_preview)
-        if link_extracted_data.HasField("link_image"):
-            _clone_collect_cf(binaries, link_extracted_data.link_image)
-
-    for field_metadata in bm.field_metadata:
-        if field_metadata.metadata.metadata.HasField("thumbnail"):
-            _clone_collect_cf(binaries, field_metadata.metadata.metadata.thumbnail)
-
-    return binaries
-
-
-def _clone_collect_cf(
-    binaries: list[resources_pb2.CloudFile], origin: resources_pb2.CloudFile
-):
-    cf = resources_pb2.CloudFile()
-    cf.CopyFrom(origin)
-    # Mark the cloud file of the broker message being exported as export source
-    # so that it's clear that is part of an export while importing.
-    origin.source = resources_pb2.CloudFile.Source.EXPORT
-    binaries.append(cf)
-
-
-async def download_binary(
-    context: ApplicationContext, cf: resources_pb2.CloudFile
-) -> AsyncGenerator[bytes, None]:
-    async for data in context.blob_storage.download(cf.bucket_name, cf.uri):
-        yield data
-
-
-async def get_entities(context: ApplicationContext, kbid: str) -> kb_pb2.EntitiesGroups:
-    edm = EntitiesDataManager(context.kv_driver)
-    return await edm.get_entities_groups(kbid)
-
-
-async def get_labels(context: ApplicationContext, kbid: str) -> kb_pb2.Labels:
-    ldm = LabelsDataManager(context.kv_driver)
-    return await ldm.get_labels(kbid)
